+++ packages/components/ui/src/index.js
@@ -1,6 +1,12 @@
+export * from './actions/resizeObserver';
+export {default as FontsLoader} from './drivers/fonts/FontsLoader.svelte';
 export * from './icons';
+export {default as FetchDriver} from './io/net/FetchDriver.svelte';
+export {default as StorageIO} from './io/storage/StorageIO.svelte';
 export * from './sensors';
 export * from './utils/env';
+export * from './utils/rx';
+export * from './utils/shared';
 export {default as Link} from './Link.svelte';
 export {default as LinkButton} from './LinkButton.svelte';
 export {default as LoadingView} from './LoadingView.svelte';
@@ -9,3 +15,5 @@ export {default as MultiBanner} from './MultiBanner.svelte';
 export {default as NoScript} from './NoScript.svelte';
 export {default as ResponsiveFlex} from './ResponsiveFlex.svelte';
 export {default as Switch} from './Switch.svelte';
+
+// TODO export a11ymenu